requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
courlan==1.3.2
dateparser==1.2.1
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-haystack==3.1.0
exceptiongroup==1.3.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
gdown==5.2.0
h11==0.16.0
haystack-ai==2.14.2
haystack-experimental==0.10.0
hf-xet==1.1.4
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
lazy_imports==1.0.0
load-dotenv==0.1.0
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-plain==1.0.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.87.0
packaging==25.0
pillow==11.2.1
posthog==5.0.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==5.6.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tld==0.13.1
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==1.26.15
yarl==1.20.1
